requirements.txt
# Core ML Libraries
numpy
pandas
scikit-learn
tensorflow
torch
transformers
xgboost
lightgbm

# Data Collection & APIs
yfinance
tweepy
requests
beautifulsoup4
feedparser

# Web Framework
flask
flask-cors
flask-socketio
gunicorn

# Database
sqlite3

# Data Processing
nltk
textblob
vaderSentiment
spacey
re

# Visualization & Analysis
matplotlib
seaborn
plotly

# Utilities
python-dotenv
schedule
datetime
threading
asyncio
websockets
eventlet

# Testing
pytest
pytest-cov

# Model Serving
joblib==1.3.1
pickle5==0.0.11
